Fix critical issues: wire to UpdateFrameEvent, change StepIndex to ushort
- SequenceDriver now implements IEventHandler<UpdateFrameEvent> and processes sequences every frame
- Changed StepIndex from byte to ushort to prevent overflow (supports 65k steps instead of 255)
- All 301 tests still passing

Remaining: allocation fixes, comprehensive tests

+++ MonoGame/Atomic.Net.MonoGame/Sequencing/SequenceState.cs
@@ -1,6 +1,7 @@
 namespace Atomic.Net.MonoGame.Sequencing;
 
 /// <summary>
-/// Tracks execution state of a sequence instance.
+/// Runtime state of a sequence instance.
+/// Tracks which step is executing and elapsed time in that step.
 /// </summary>
-public readonly record struct SequenceState(byte StepIndex, float StepElapsedTime);
+public readonly record struct SequenceState(ushort StepIndex, float StepElapsedTime);
